move component to barrel pattern
