+++ src/components/index.js
@@ -1,0 +1,9 @@
+export { default as Box } from './Box'
+export { default as Button } from './Button'
+export { default as Grid } from './Grid'
+export { default as Header } from './Header'
+export { default as Input } from './Input'
+export { default as Layout } from './Layout'
+export { default as Spacer } from './Spacer'
+export { default as Stack } from './Stack'
+export * from './typography'
