+++ src/templates/Page.jsx
@@ -1,10 +1,6 @@
 import React from 'react'
 import { graphql } from 'gatsby'
-import Layout from '../components/Layout'
-import Header from '../components/Header'
-import Stack from '../components/layout/Stack'
-import Spacer from '../components/layout/Spacer'
-import Grid from '../components/layout/Grid'
+import { Layout, Header, Stack, Spacer, Grid } from '../components'
 
 const Page = ({ data }) => {
   if (!data) return null
